feedback

+++ terrazzo-api/src/controllers/membershipController.ts
@@ -3,6 +3,7 @@ import { createOrganizationMembershipDb, deleteOrganizationMembershipDb, getOrga
 import { getOrgByIdDb } from '@trz-api/persistence/organizationPersistence';
 import { setRoleIdsForUserInOrgDb } from '@trz-api/persistence/roleAssignmentPersistence';
 import { getUserHeaderByIdDb } from '@trz-api/persistence/userPersistence';
+import { userIsOrgOwner } from './organizationController';
 
 export const getMembersInOrg = async (orgId: OrganizationId) => {
     const org = await getOrgByIdDb(orgId);
@@ -47,6 +48,10 @@ export const createMembershipIfDoesntExist = async (membershipRecord: Membership
 };
 
 export const removeMembership = async (userId: UserId, orgId: OrganizationId) => {
+    const isOwner = await userIsOrgOwner(userId, orgId);
+    if (isOwner) {
+        throw new Error('Cannot remove membership for the owner of the organization');
+    }
     await deleteOrganizationMembershipDb(userId, orgId);
     await setRoleIdsForUserInOrgDb(userId, orgId, []);
 };
